moved chips, mult and retrigger to round.c
